Can get more flexibility from intercepts_on_line by making it using f64s (eg, can resample a contour very easily this way, or use intercepts_on_line to approximate a distanc field directly)
Might need to put the correction code in the edge cell interator though

+++ src/bezier/rasterize/path_contour.rs
@@ -90,7 +90,7 @@ impl<'a> SampledContour for &'a PathContour {
     }
 
     #[inline]
-    fn intercepts_on_line(self, y: usize) -> SmallVec<[Range<usize>; 4]> {
+    fn intercepts_on_line(self, y: f64) -> SmallVec<[Range<f64>; 4]> {
         (&self.contour).intercepts_on_line(y)
     }
 }
